% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{plot_alignment_hist}
\alias{plot_alignment_hist}
\title{Create and return alignment length histogram (ggplot object)}
\usage{
plot_alignment_hist(blast, rep_seqs, cutoff_fraction = 0.6)
}
\arguments{
\item{blast}{BLAST data frame.}

\item{rep_seqs}{Named list/character vector of DNA strings (from seqinr::read.fasta(as.string = TRUE)).}

\item{cutoff_fraction}{Numeric; fraction of median alignment length for cutoff line. Default 0.6.}
}
\value{
A ggplot object.
}
\description{
Create and return alignment length histogram (ggplot object)
}
