% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{trim_alignments}
\alias{trim_alignments}
\title{Trim BLAST alignments by minimum length}
\usage{
trim_alignments(blast, rep_seqs, fraction = 0.6)
}
\arguments{
\item{blast}{BLAST data frame.}

\item{rep_seqs}{Named list/character vector of DNA strings (from seqinr::read.fasta(as.string = TRUE)).}

\item{fraction}{Numeric; fraction of the median rep-seq length used as the cutoff. Default 0.6.}
}
\value{
Filtered BLAST data frame.
}
\description{
Trim BLAST alignments by minimum length
}
