% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{plot_alignment_hist}
\alias{plot_alignment_hist}
\title{Create and return alignment length histogram (ggplot object)}
\usage{
plot_alignment_hist(blast, rep_seqs, cutoff_fraction = 0.6)
}
\arguments{
\item{blast}{BLAST dataframe.}

\item{rep_seqs}{Named list of DNA strings (from \code{seqinr::read.fasta(as.string = TRUE)}).}

\item{cutoff_fraction}{Numeric, fraction of median for cutoff (default: 0.6).}
}
\value{
ggplot object (for later use in PDF or display).
}
\description{
Create and return alignment length histogram (ggplot object)
}
